Update patch set 2
Patch Set 2:

I think it would easy to build a follow-on change to split this into a separate server process.  I think that will be compatible with my future plans for the nodepool daemon as well (if we go fully zookeeper).  I don't have time ATM to write that, but if someone else wants to, I'm not opposed to the approach.

(I will note that I didn't think it was immediately necessary because the process will require database access which likely means file-level access to the credentials unless we give it its own unix user.  Also this code is very simple and any escape bugs are likely to also be present in a second-process implementation.)

Patch-set: 2
Label: Verified=0