Update patch set 2
Abandoned

Re reading the documentation, 'name-filter' is probably documented to be filtering against the flavor name:

 Additional filter complementing min-ram, will be required to match on the flavor-name 

So it is all good if you read the doc properly :-}  I think my original concern was the example stating 'performance', it is now 'something to match' which is referred to in the explanation paragraph.

So this change is no more needed.

Patch-set: 2
Status: abandoned