Cleanup building dib images on startup
When nodepool restarts the image build worker detects this and stops
building any images that were building. This means that nodepoold needs
to clean out any building images on start otherwise it will be out of
sync with the builder. Perform this cleanup by setting the dib image
state to delete if it was in a building state on startup. The periodic
cleanup will then delete it on the next periodic cleanup.

Change-Id: I27df8d11e70b7651f82c84e8e6ffb08ae3e3ebdd