Update patch set 2
Patch Set 2:

The 255 character limit for a metadata entry makes this worth considering.  I will note that there is another limit that hasn't been mentioned: the number of metadata entries available.  A survey of the clouds we use indicates that value ranges between 20 and 128.  This patch will mean we are less likely to hit the 255 char limit on a single metadata entry, however, it does bring total metadata consumption from 2 to 5 entries.

Patch-set: 2
Label: Verified=0