Update patch set 1
Patch Set 1: Code-Review-1

This assumes that nodepool is the only owner of floating ips. If I am using the same account/project to also manage static instances, this may not be the case, so I would prefer to at least have an option to disable the cleanup.

Patch-set: 1
Label: Code-Review=-1