Add namespace for nodepool
If multiple nodepools share a single tenant, one nodepool's Servers
may be deleted by other nodepools' cleanupLeakedInstances(). So add
namespace to specify each nodepool's scope. Each nodepool should only
clean Servers which are marked belong to its namespace. If the namespace
is not set in the configure file, by default use the hostname as the
namespace.

Change-Id: I430e2989c1c31acab4e8b3f390b883ae913e5711