Update patch set 2
Patch Set 2: Code-Review-1

I'm generally against the migration path at this point since we have an initiative to move us towards zookeeper for the state data in nodepool. If we are going to continue to support both SQL and ZK, then this makes sense.

Knowing the planned direction will go a long way towards knowing if this is the right direction. This also will need a rebase/some work before it could land.

Patch-set: 2
Label: Code-Review=-1