Update patch set 13
Patch Set 13: Code-Review-1

This encodes a behavior that is not actually what we want to solve the problem.  It would be simpler to just immediately upload images for a diskimage as soon as it has completed building.  That will similarly alter the upload behavior to what you want, eliminate an unecessary delay for all images to complete their builds, and leave us with less code (and future tech debt in the form of a test that insists on a side effect of the main thing we're trying to accomplish here).

Patch-set: 13
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1