Update patch set 4
Patch Set 4: Code-Review-1

After confirming with clarkb we create sparse images, i'm changing my vote to a -1 since this will not reflect actual usage (which seems to be the intent).

Feel free to override my -1 if the intent is to cover the actual file size as represented (maximum) vs. the "used space" in the case of sparse files.

Patch-set: 4
Label: Code-Review=-1