Python 3 Fixes: Use of bytes as appropriate
Use bytes where appropriate instead of str due to the changes in
handling of bytes and strings in python 3. Notably gear protocol
requires everything to be bytes on the wire so the jobs are
handling most of the conversion to/from string to bytes via explicit
utf-8 encoding/decoding.

StringIO in the fakeprovider is replaced by six.StringIO due to
the move to the `io` module.

Python 3 tests will continue to fail until a fix to gear library is
released to eliminate a change of dictionary while iterating. See
https://review.openstack.org/#/c/322352/ for more information.

Change-Id: I94073fda5796ad277ea55ee1ba149810c1c9b8fd