Python 3 Fix: round() behaves differently in py3
In python 3 round() built-in will always round down "half" instead
of the python 2 behavior of rounding up. To work around this so that
nodepool allocations continue to behave "correctly" (established
behavior), instead of relying on round(),
`nodepool.allocation.AllocationProvider.make_grants()` now utilizes
the `decimal` module and performs a `quantize()` with rounding
explicitly defined to ROUND_HALF_UP. The `Decimal` result is casted
to an explicit int to ensure consistent behaviors.

Change-Id: Id18e40493e25bf66c04a82ce3a18f411e4a933f1