WIP - Python 3 Fix: Gearman Server not starting with fixture
A second call to the gearman server fixture is not properly starting
a new gearman server under PY34. This is also skipping the test that
is having issues with the gear library due to a use of .values() from
a dict where the dict is changed as part of a loop (in py3 .values is
an ITER not a rendered list).

Change-Id: Ie27a5013bfe72a257ae1eb52c73514a0f02f5f6c