Update patch set 2
Patch Set 2: Code-Review-1

I think this patch is not about vote on removing snapshot code from nodepool. It is about deprecating it and adding a warning when snapshot code is used. As such, I am +1 for better deprecation process.

Regarding removing snapshot code from nodepool (which this patch is not doing), I'd like there to be a bit of time before it is removed: it creates a lot of work for people who still use snapshots. We can pin nodepool for awhile, but there would be a point when it would stop working with other parts of the CI (Zuul v3?). While dib is overall better approach, it adds restrictions that snapshot image process does not have (e.g. I could not get it to build across architectures). Yes, you do not need nodepool to build snapshot images. I can think of a way to implement snapshot image builds in bash. But it would need time investment, maybe more so than hacking around dib restrictions.

Patch-set: 2