Update patch set 1
Patch Set 1: Code-Review-1

This seems generally correct, though I think we should switch the value of the job_worker_map to a set rather than a list since, as currently written, if you do register jobs with labels, you will end up with duplicate entries.

As a set, we should be able to safely populate it twice.

Patch-set: 1
Label: Code-Review=-1