Update patch set 2
Patch Set 2: Code-Review-1

In the MySQL schema fixture, we concatenate os.getpid() to the random string used for the per-test-runner database (so it looks like "randomstuff_pid"), since we know each test runner is in a different process.  If you do the same here, we should guarantee no collisions.

Patch-set: 2
Label: Code-Review=-1