Better handling of gearman status and workers
There were two major issues with the old demand calculation code based
on the status and workers output of gearman. First we tried to lookup
functions even though that had 0 workers which we know won't work so
skip these. Second we were not handling static workers that aren't
managed by nodepool.

Handling this second item is a bit trickier. We update the code to have
3 states for the function worker map.
A) Function known and workers identified
B) Function known and workers not identified (static external workers)
C) Function not known at all

We will only trigger the rebuilding of the map for state C or a ttl
timeout.

Change-Id: I485c52eb992c77476b02ac3fceeaa8ddac2b724e