Update patch set 1
Abandoned

We actually don't think the overhead of the additional job registrations using the old system will be that drastic. This code has become quite complicated and we still haven't gotten it right and 'workers' is not very cheap either.

All this to say that we should stick with the old method and this idea though neat is not panning out.

Patch-set: 1
Status: abandoned