Update patch set 4
Patch Set 4:

(2 comments)

Sorry I forgot about this change.   Will amend to copy part of the commit message to the inline doc for discoverability, good suggestion.

For 'nodepoold' section, this patch only add a setting for the deletion delay.  Part of the intent is to add all the other hardcoded settings so Nodepoold can be finely tuned.  Should I parameterized the rest of them ?

Patch-set: 4
Label: Verified=0