Update patch set 1
Patch Set 1:

We have separate build and upload workers because building an image and uploading have very different performance requirements (one is cpu/io heavy, the other is largely network dependent).  In practice, we would likely size the builder host for one build worker, but it's configurable in case someone is running on beefy hardware.  The real reason the numbers of workers are configurable is for the upload workers, where even if we size a host for one builder, we have found that we can sustain 16 simultaneous image uploads.

After building one image (really one type of one image), we then upload that same image file to all of the cloud providers that share that type.  So we create one qcow for ubuntu-xenial, and then upload it to, say, 5 clouds.

Also, the current version of the nodepool builder does not handle live configuration changes well.  We somewhat frequently adjust our image configuration (adding or retiring an image), and currently the builder requires a restart.  That's an oversight -- we don't think that should be necessary.

All of that to say that while we could put a limit on the number of build workers (it would actually be the number of image files needed, which is a function of not only the number of images, but also the file types as dictated by the clouds to which we are uploading them), I think it might be better to omit that and just let the sysadmin specify it based on the system capacity.  The number will change as images are added and removed, or existing images are added and removed to/from clouds.  And eventually, we want that to happen without a restart.  The useful limit of upload workers is the sum of the number images used in each cloud (in our case, that approximates images*clouds).  Again, I'd suggest just going with whatever the admin configured, and not worry about having idle workers.

It does seem that the worker model may still have relevance here -- even though it was *mostly* put in to map to gearman concepts, it seems that especially if having multiple build and upload workers in a single process is desirable (I think so -- especially upload), then it's still a useful way to think about things.  We could also have a master/subprocess model instead of using threads internally.  I think the main thing is that it is nice to have a very simple dedicated thread of execution for each of these things, but also the ability to easily start more than one of those with a simple 'service nodepool-builder start', whether that's multiple threads inside one process as now, or multiple processes started by a single overseer.  IOW, feel free to change it, but if what's there works for you, it works for me.

But I don't think I'd recommend assigning specific images to workers.  We *do* need to inform the builder which images it is capable of building (for instance, a host that can only build ARM images), but I think it will be easier to deal with internal logic where a build worker inside of a builder process can build any image that the builder process is configured to be able to build.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0