WIP Fetch server console log if ssh connection fails
Currently, if the ssh connection fails, we are blind to what the
possible failures are.  As a result, attempt to fetch the server
console log to help debug the failure.

TEMPORARY: add a raw parameter to submitTask, because get_server_console
needs it for now (we can make that go away, but not until we can land
patches for shade) - do not land this with the raw parameter still in
place.

Change-Id: I39ec1fe591d6602a3d494ac79ffa6d2203b5676b
Signed-off-by: Paul Belanger <pabelanger@redhat.com>