Update patch set 1
Patch Set 1: Code-Review-1

A soft -1 on this -- if the ready script fails, I would think that the *output of the ready script* rather than the console log would be the most important thing.  And I think we do log the ready script output.  So I'm confused by the statement that we are "blind" to failures.

Patch-set: 1
Label: Code-Review=-1