Update patch set 1
Patch Set 1:

(1 comment)

Jeremy suggested in IRC that this might be useful in the case where we could not connect to run the ready script.  That makes more sense to me.  I think I would rather see a patch that only retrieved the console log on connection failure.  If we do that, then I don't think we need to add the option -- I think we can just always perform the collection.

Patch-set: 1