Use interface_ip instead of internal logic
shade has auto-detection of which of the ip's in a server are usable.
Most notably, it detects if the host can handle IPv6 and if so it
defaults to using it. Skip doing v6 checks directly in nodepool and rely
on shade's interface_ip value.

This changes behavior on clouds with IPv6 available but not explicitly
configured as being "ipv6_preferred" to using the IPv6 address and not
the IPv4.

Change-Id: Id70b66a5f05dce34066d86f7c86b6e85a4ee31aa