Plumb ipv6-preferred into force-ipv4 setting
We collect an ipv6-preferred setting, but we don't pass it to shade.
This is currently a problem for devstack nodepool tests with the new
support for dual stack networks, because devstack has an OSIC-like
dual-stack network. For the moment, we need to force_ipv4 so that the
tests will work right. In the next patch, we'll fix nodepool to use ipv6
if detected (which shade can do) but we can't turn that on until IPv6 is
fixed in Rackspace.

Change-Id: If356106dad846801393e19f97f51498b10d381be