Use priority queues for worker threads
Nodepool builder is a classic producer-consumer program. Requests
for builds or uploads are being produced by the main thread, and
the workers will consume those requests. We use priority queues
so we can put special messages at the front of the queue to terminate
the workers, effectively ignoring any unprocessed requests.

Change-Id: Id5da0431c91d0fce749a6ff1a5419f694c1f5d6c