Update patch set 1
Patch Set 1: Code-Review-1

I don't think we need queueing for this -- the idea in the spec is that an individual builder can decide on its own that an image needs to be built, and then act upon that.  This holds true for both multiple build threads on a single host as well as multiple builder hosts.  Our production system is likely to have more than one build host, with one build thread on each host -- I try to think about that scenario when I imagine this.

We don't want a builder to accept more work than it can handle -- instead, we want that work to "back up" into Zookeeper so that other builders see work that they can do and do it.  Further, we don't need a queue in Zookeeper either, because this problem doesn't actually require centralized scheduling.  It's basically a state machine that, left unchecked, drifts naturally out of its target parameters.  The builders continually look at the state, and when they detect a part of it out of whack, act to correct it.  Things that put it out of whack are:

* Time has advanced and the last build time for an image is before the most recent scheduled build time for that image as specified in the config file (the image build "cron").  The builder should try to build it.
* A new image has appeared in the config.  The builder should try to build it.
* This builder has built images on disk, but some of the providers don't yet have an uploaded copy of those images.  The builder should upload it.
* This builder has an image on disk whose state is set to delete.  The builder should delete it.
* There is an uploaded image whose state is set to delete.  The builder should delete it.

There are probably others, but this illustrates what I mean about the logic being in the edge.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1