Unsubvert image and flavor caching
Recent shade allows users to pass in image and flavor to create_server
by name. This results in a potential extra lookup to find the image and
flavor. Since nodepool is not using shade caching, this is causing our
nodepool-level caching to be subverted. Although an eventual project to
get nodepool to use shade caching, that's a bad scope creep for now - so
just pass in the ids (which we know are ids) in dicts, which gets shade
to not attempt to look for them.

Change-Id: Ic70b19ad5baf25413e20a658163ca718dce63bee