Update patch set 3
Patch Set 3:

I used to really like subclassing threads, and then some folks suggested that maybe not subclassing is the way to go, especially considering potential collisions with some of the attributes on threading.Thread.

So recently I switched to the 'has-a' rather than 'is-a' model.  I make classes that "have-a" thread.  They still behave somewhat like threads -- they have start/stop methods which then launch or join their internal thread.  But the implementation of that is hidden from use.  I also think it's nicer in that you're presenting a simpler API -- "here is a worker, you can start and stop it" without presenting the whole Thread API.  I think you could do the simplication you are doing here but still not actually subclass Thread.

Anyway, this isn't that important (it's even pretty easy to change -- we have start/stop methods either way), but I thought I'd bring it up in case we wanted to align on one way or the other.

Patch-set: 3