Use MySQL create user statement to silence warning
The following warning shows up in unit test runs:

  Warning: (1287, 'Using GRANT for creating new user is deprecated and
  will be removed in future release. Create new user with CREATE USER
  statement.')

It's not a problem - but it is test noise, so remove it.

Also, remove the use of flush privileges. It is not needed when GRANT
statements are used. It's only necessary when the user tables in the
mysql schema have been modified directly.

Change-Id: I9da567725ebffba42d22d2236bbf2c84ebe1f1fe