Add skeleton for DIB build cleanup
This adds two portions:

1) Adds a periodic check for marking expired builds for deletion. These
would be builds that are 'ready' and not the current or previous build,
and older than 8 hours (matching current logic). Also marks failed
builds for deletion that are older than 8 hours.

2) Adds a periodic check for performing the deletion of the local DIB
builds marked as needing deleted, as well as for removing build data
in ZooKeeper that is irrelevant (unused or junk nodes).

Also cleans up some docstrings for ZK API.

Change-Id: I01ea82937664ae4cb0b80b7efe85a19ad18d7f69