Use sequence znodes for image build numbers
ZooKeeper has a concept of sequence nodes[1] that we can use rather
than our current method of querying for max used build number. This has
advantages: Current logic is flawed when the build number rolls over
at MAXINT (we would keep trying to create a znode with the same build
number). It also eliminates some ZK API calls.

Because we now have the getMostRecentBuild() API method, the
getMaxBuildId() API method is unneeded.

The logic for upload IDs will be updated similarly in a separate
review.

[1] https://zookeeper.apache.org/doc/trunk/zookeeperProgrammers.html#Sequence+Nodes+--+Unique+Naming

Change-Id: I34e1b6349b2c1cb07dce0c2e154ada8804c7e588