Handle config reload race
Nodepool code reloads its config periodically to dynamically
reconfigure itself. If something updates the config, causing it
to temporarily not be available (e.g., saving from vi), it's
possible it won't exist if we try to read from it at the same
time. Add a simple retry to the config loader.

Change-Id: I4a8887130e4d2d667354a169e04c59177bd8ace9