Increase build worker main loop delay
This doesn't need to run very often (it's triggering things that
under normal circumstances happen once per day and take hours to
run).  In the case of a manual build request, the operator can
probably wait 10 seconds.

Also, remove the TODO that says it should be configurable -- this
probably isn't actually very interesting for an operator to change.

Change-Id: I6f0e6a13d82d3b375cc4a749a69a5d2d52c84d11