Change getMostRecentBuild to get multiple builds
The getMostRecentBuild() method current returns a single build
that is considered the most recent (by state_time). In a future
change, we will need to be able to find X number of most recent
builds when we begin considering which ones to delete/cleanup.
This change allows for a new parameter to specify the count of
builds to return that have the latest state_time. The method
name is adjusted appropriately, too, to be plural.

Change-Id: I754cb065334afe48bd2ef5bc40667c3166b95bb9