Add local image cleanup to BuildWorker
Remove any local DIB image builds that have aged out.

Note that this does not yet remove the image uploaded to the
providers. The build is marked as 'deleted' in ZooKeeper before
it is deleted from disk, so the next stage is to modify the
UploadWorkers to look for these 'deleted' builds and remove
them from the provider, then remove the build znode in ZK.

Change-Id: If0beca3b1da19537cff1d4e675d5245c721ce2a2