Update patch set 2
Patch Set 2: Code-Review-1

I agree with this change, but the other getMostRecent* methods have changed (or hopefully WILL change when the reviews merge) since you originally wrote this. They all now take a count, return a list, and can ignore state (see review 385590).

I actually need something similar to this code for review 387583. I have a facsimile of it in the _buildUploadRecencyTable() method of that change that you could steal.

Patch-set: 2
Label: Code-Review=-1