Update patch set 1
Patch Set 1: Code-Review+2

I notice that you don't actually take advantage of this in the way I expected from the commit message -- code that was previously:

  if builds is None:

is translated to

  if not builds:

Which isn't substantially different.  You could remove the check entirely, since the callers of this method handle the empty list case without alteration.

It's fine either way -- I'm mostly wondering what the impetus was, and where you want to go in the long run.

Patch-set: 1
Label: Code-Review=+2