Update patch set 1
Patch Set 1:

Do we need to use iteritems() as opposed to items() (which is supported in both versions of the language)?

AIUI, one reason not to use items in py2 is that it computes the entire list; our lists will be small enough that the additional ram is not a concern.

Another difference lies in whether they copy the dictionary or whether insertions or deletions are reflected during iteration.

Patch-set: 1