Update patch set 2
Patch Set 2: Code-Review+2

> I believe this change enforces running nodepool in the same arch as
 > the target image runs. Non-x86 3rd-party CI players would break.

This isn't entirely the case. First as mentioned in other comments this is just a warning at this point. Snapshot builds will continue to function. But also DIB should support builds for power and arm (I am told that people are using it for this). The caveat being that it doesn't do cross builds so you have to run your nodepool builder on the target arch.

Now is probably a good time to start testing this and file bugs/fixes against DIB as necessary to make this work for you. I am sure the DIB team would be happy to help as well.

Patch-set: 2
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2