WIP: Transition ZK API from dict to object model
The methods to get and store builds and uploads will now take and/or
return objects of ImageBuild or ImageUpload types.

ZooKeeper API methods that used to return tuples (to include a build
or upload ID along with the corresponding data) are now simplified to
return a list of objects which will contain the ID.

Since our object model validates state values, we no longer need to
test for empty state values (which are not allowed), so two tests
accounting for that are removed.

Story: 2000767

Change-Id: I4e97f8d0d1e3095a6c3e988a8266805a82a0b520