Separate image upload logs into separate logger
By making the image upload logs go into a separate logger, we can add
filters to put them into separate logs files in the same way we do for
image builds.  This will help debug upload issues with specific
providers.

Small changes to the logging is made to use the provider logger in the
failure exception case, to make sure the error message gets into the
upload logs.

Nothing will really change until the log-config file is updated to
filter these out; see Ic80e532891f039e0f835847ddfbae38f1ce3cd8a and
corresponding puppet changes.

Another option would be to append this info to the builder log.
That's kind of logical (see the build, see the uploads).  However, I
was not sufficiently convinced that a) we're 100% sure the output is
clean enough to export to the web and b) that we really want exception
failures for providers publically.  With this, we can easily publish
the upload logs if required, but they remain unexported for now.

Change-Id: Id58d5c9fd9f8c6bab2e1ee05bedada22ef7e08a0