Update patch set 6
Patch Set 6: Code-Review+1

I'm not a fan of the long and misleading zookeeper_client variable name (it's not a zk client, and it's long to type... 'zk' would be more consistent with builder code), but not worth a -1 for that. We can change it later.

Patch-set: 6
Reviewer: Gerrit User 3099 <3099@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1