Add check for uploads to deleteBuild()
Deleting a build while we still have provider upload znodes is a
no-no, which is why we didn't recursively delete from the build
node. However, we have lock nodes remaining that aren't currently
cleaned up. Add a check to deleteBuild() that verifies if any
upload nodes remain, and have it return True/False based on whether
or not it could successfully delete the build node. Also, we can
now safely recursively delete to get the left over lock nodes.

Change-Id: I782f3e252dc296efbaa24da6ac1672f4752233a8