Allow configuring path to disk-image-create
I'm in a situation where I'm trying to use a disk-image-create
application that is not in the PATH. There are ways I can hack this into
our environment however it would make things easier to configure
nodepool to point to a different executable.

Adding this configuration allows us to use the config to set the
fake-image-create app in tests rather than have to hardcode a check for
fake- in the image name.

Change-Id: Ia2963a92e61ad00c72de8ade38322dc899c0a441