Update patch set 4
Patch Set 4: Code-Review+2

So, I understand this is the only way to do it today. But this is much more complicated then how we do it today.  I believe the intent is to review this, but +2 for code coverage right now.

Patch-set: 4
Reviewer: Gerrit User 4162 <4162@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2