Update patch set 6
Patch Set 6: Code-Review-1

I actually think that whether an image is mentioned in a label should not affect anything about whether the image is built or uploaded.  Instead, it should only be used to map images to nodes.

We can use the presence of an image directly on a provider to indicate whether it should be uploaded, and we can use the presence of a diskimage stanza at all to indicate whether it should be built.

Patch-set: 6
Label: Code-Review=-1