Delete builds when diskimage removed from config
We want to remove an image from the system entirely when we remove
that image from the diskimage section of the config.  This means
that we should be able to remove the image from zk and providers
iff the image is found locally on this builder.  If it is not,
we assume that some other builder is responsible for it.

Change-Id: I616e1f9771cbda93bc68166534b3a73fd4ed94b8