Re-enable alien-image-list command and tests
Changes the alien-image-list command to compare provider images
and those recorded in ZooKeeper.

Re-enables the test_alient_image_list_fail test and adds a new test
(test_alien_image_list_empty) to verify normal operation.

NOTE: I'm unclear as to how to artificially create an alien image
in FakeProvider to test the "alien image found" case, but the
pathway is there to add a test for it if that is, indeed, possible.

Change-Id: Ib68e05edccd0f7aac10593edae0a7098a6f30de0