Add pause support for diskimages
There are times where you want to disable diskimage builds happening
by nodepool-builder. For example, if diskimage-builder is producing
broken images.  In this example, we'd want to pause image builds and
roll back to a good image.

This prevents nodepool-builder from replacing the good image with a
potentially bad one.

Change-Id: Ib39e1914598a8007cccc31e261099dd7565b7527
Signed-off-by: Paul Belanger <pabelanger@redhat.com>