Don't use taskmanagers in builder
ProviderManager is a TaskManager, and TaskManagers are intended
to serialize API requests to a single cloud from multiple threads.
Currently each worker in the builder has its own set of
ProviderManagers.  That means that we are performing cloud API calls
in parallel.  That's probably okay since we perform very few of them,
mostly image uploads and deletes.  And in fact, we probably want
to avoid blocking on image uploads.

However, there is a thread associated with each of these
ProviderManagers, and even though they are idle, in aggregate they
add up to a significant CPU cost.

This makes the use of a TaskManager by a ProviderManager optional
and sets the builder not to use it in order to avoid spawning these
useless threads.

Change-Id: Iaf6498c34a38c384b85d3ab568c43dab0bcdd3d5