Make sure we clean up diskimages with invalid external_name
It is possible for external_name on our upload images to be empty.
Either we have written the data to zookeeper incorrect or some other
unknown.

As a result, check if external_name is set, then delete the image.
This will allow our cleanup worker to properly delete unneeded
diskimages from the HDD.

Change-Id: I2bad5f2c2585c54879497de3c8997dd6234be5a7
Signed-off-by: Paul Belanger <pabelanger@redhat.com>