Update patch set 11
Patch Set 11:

tl;dr  I agree with Jim that PS10 is better.

So turns out my idea really isn't any better than what was in patchset 10, which I think is the better patch. I think we could definitely improve this later by totally reworking the DibImageFile class, which was never meant to support operations on the files, but rather act as a convenient container for the pieces making up the file on-disk. Doing that here is asking too much, I think.

Patch-set: 11