WIP: Remove ZK lock variables
DO NOT MERGE / WIP

In anticipation of reducing the number of ZK client connections that
are used, we need to begin adjusting our ZK API for use by multiple
threads. This begins removing the ZK lock variables. The tests that
need multiple ZK connections (locking tests) had to be modified to
use a separate thread for the test.

Note that right now separate clients are still created for each new
ZooKeeper object. That will eventually change.

Also, remove the ZK API call for registering watches since that
entire thing needs reconsideration for this new structure (and it
wasn't used anyway).

Change-Id: Ib0bc2157a1b6fbbca1d12c25a1f8d439143480ba