Use is_public to determine vendor images
Images uploaded by cloud vendors are marked "is_public", and marking an
image public is not allowed to normal users of the clouds. Use that to
filter out images that the cloud vendor uploaded.

Add a test that tests that alien images are listed. Also add a vendor
image to both the empty test and the alien test to verify that vendor
images are not listed.

Change-Id: Idd5e98fb5303ec27dc0727ef65e2c51c0097dadb