Update patch set 1
Patch Set 1: Code-Review-1

First, I don't know that we want to vomit a traceback to the console if there's a network hiccup.  I think changing the exception to a warning was there for a reason.

If you *did* want to do that (please don't without further discussion), just remove the exception handler.  The try/except/raise that this patch does only serves to hide the actual exception and traceback (which might be useful) with a useless traceback.

If we wanted to test that an error was logged, we can set up a log handler and check the output in the test.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1