Stop using FakeProviderManager
We also monkeypatch ProviderManager in test cases. Having both versions
of this leads to one code path which gets one copy of the cloud object
and one code path which gets a whole different object.

However, FakeProviderManager is useful for running nodepool "for rea"
but without talking to clouds. Keep it around for that.

Change-Id: Idb4b8af8972d9ab29abf6e7da4f9ef4d073204ee