Consume Task and TaskManager from openstacksdk
openstacksdk's TaskManager has had to grow some features to deal with
SwiftService. Making nodepool's TaskManager a subclass of openstacksdk's
TaskManager ensures that we get the thread pool set up properly.

openstacksdk has the same Task class as nodepool, so there's no good reason to
keep a second copy of it. Also, we don't actually use the Task class
directly any more.

Turn TaskManager from an is-a Thread to a has-a thread, because
otherwise multi-inheritance and constructors get weird.

Finally, remove the client parameter from TaskManager which is a no-op
and not present in the base class.

Change-Id: I241bd6a2ebdfdaf83f11612b2addb002b7f55fb4