Consume Task and TaskManager from shade
shade's TaskManager has had to grow some features to deal with
SwiftService. Making nodepool's TaskManager a subclass of shade's
TaskManager ensures that we get the thread pool set up properly. Also,
shade has the same Task class as nodepool, so there's no good reason to
keep a second copy of it.

Turn TaskManager from an is-a Thread to a has-a thread, because
otherwise multi-inheritance and constructors get weird. That results in
just having the TaskManager stop method call the thread join instead of
having TaskManager have a join and a stop method itself.

Change-Id: I241bd6a2ebdfdaf83f11612b2addb002b7f55fb4
Depends-On: Ib48fe54feb8b6ed10da3e5fa9accf985e4b22623