Consume Task and TaskManager from openstacksdk
openstacksdk's TaskManager has had to grow some features to deal with
SwiftService. Making nodepool's TaskManager a subclass of openstacksdk's
TaskManager ensures that we get the thread pool set up properly.

Finally, remove the client parameter from TaskManager which is a no-op
and not present in the base class. We also don't need the setClient
call which is also a noop.

Change-Id: I241bd6a2ebdfdaf83f11612b2addb002b7f55fb4