Update patch set 1
Patch Set 1: Code-Review+2

(4 comments)

I highlighted two chunks of code which probably still have an analog in the new world.  They aren't rocket science, and we can always come back to this if we need it.  I'm fine with either just remembering that we can dig into the history for it if we need it and approving this, or respinning.

Patch-set: 1
Label: Code-Review=+2