Handle provider removal from config
If a provider is removed from the config file, then that provider should
stop handling node requests. The provider thread will shut itself down
in that case.

Also, if a previously present provider is removed and then re-added back
to the config at some point, we should recognize that and restart the
thread for it.

Change-Id: I8eff48f35633d92a993c5979017bd06e7480c35a