Update patch set 1
Patch Set 1:

Paul Belanger wrote:

> If we go down this path, can we instead switch from --fake to --dib-
> cmd /--builder-cmd (your choice)? This give us the ability to remove
> the hard coded fake, and provide full path options for diskimage-
> builder (if installed into a virtualenv)

In Ia2963a92e61ad00c72de8ade38322dc899c0a441 it seemed the consensus was that we would likely move to the python API in an upcoming version of dib.  I chose this approach so that we would not introduce an option (dib path) which would need to be immediately deprecated.  Instead, --fake works with the current way of invoking dib and presumably the new way.

Patch-set: 1