Manage providers in ProviderWorker threads
We don't want every provider manager to be available in the config
file. Instead, let each thread responsible for the provider have a
single ProviderManager that it manages itself.

Since this manager is among several attributes that need to be shared
with the child NodeRequestWorker threads, let's just pass the parent
down to the children and let them grab the attributes needed.

Shade integration tests were changed to not require access to
nodepool internals and test the actual thing they care about, the
ProviderManager.

Change-Id: I1533d53ff2cdf7ca51b7e8cc96ba55a3ced1a96c