Replace provider template-hostname with hostname
We will no longer get the hostname format from the 'target' section
of the config file, and instead just grab it from the 'provider'
section. Move 'hostname' to 'provider' section and remove
'template-hostname' since that is redundant.

Note: Docs are updated, and references still remain in some of the
code that is about to be replaced (e.g., in OLDNodeLauncher class).
New config value will be used in upcoming reviews.

Change-Id: I1ab2644d31eb371e1cf6b5e67ed2d2f850cf2464