Update patch set 10
Patch Set 10: Code-Review-1

(2 comments)

There's at least one additional thing that will make this a bit more complicated, and that is availability zones.  When we get a node request from zuul, all of the nodes for it should be filled from the same AZ (whether we need to create them or they come from the pool).  When we create nodes to satisfy min-ready, we should be careful that we don't accidentally create all of them in the same AZ.

Having said all of that, I think we should gloss over that for now and implement those changes in a future change.  This looks very much like what we generally want, and it's a big enough change already.

-1: one comment inline that should be fixed.

Patch-set: 10
Label: Code-Review=-1