Add a HTML status page
This adds a sample HTML status page to display the JSON output
provided by the nodepool webapp.

It has external CDN dependencies for common bootstrap and jquery
parts.  People get touchy on this sort of thing and I don't think it
would be time well spent for us to make the nodepool webapp serve this
directly and have to deal with shipping dependencies, etc.  Although I
plan to puppet this onto nodepool.o.o it is not 100% generic with some
hard-coded paths, and again probably not worth big efforts to make it
so.  This is a long-winded way to say I put it in a contrib/
directory.

Since this works OK, I plan to add the "image-list" output too in a
follow-on change.

Change-Id: Ia0f469a93d2417505dc0a1f69c1b81ceff6ec4fb