Cleanup node request locks
Node requests are created in /nodepool/requests. The request locks are
created in /nodepool/requests-lock. When a node request is deleted, or
if it simply disappears (it is an ephemeral znode, so it could just go
away), the lock will still remain. Add code to delete locks where there
is no equivalent request.

Change-Id: I7c442bf13f0995e0361745a5817a3e244947407b