Create provider 'launcher-id' setting
Currently, while testing zuulv3, we are wanting to share the
infracloud-chocolate provider between 2 nodepool servers.  The current
issue is, if we launch nodes from zuulv3-dev.o.o, nodepool.o.o will
detect the nodes as leaked and delete them.

A way to solve this, is to create a per provider 'launcher-id' where
an admin can configure 2 separate nodepool servers to share the same
tenant.  The big reason for doing this, is so we don't have to stand
up a duplicate nodepool-builder and upload duplicate images.

Change-Id: I03a95ce7b8bf06199de7f46fd3d0f82407bec8f5
Signed-off-by: Paul Belanger <pabelanger@redhat.com>