Group nodes by availability zone
An AZ is chosen at random when selecting a node set for a node request.
If the AZs are not listed within the config, then we cannot guarantee
AZ grouping.

To avoid selecting an AZ that may not match any nodes within our READY
pool (because we ALWAYS want to use READY nodes when they're available),
we use the AZ from the first READY node for the random AZ choice. If we
bypass the READY pool (e.g, min-ready requests), or if the READY pool is
depleted, then we randomly select an AZ from the config.

Change-Id: I97bdda2c4b6952f1cea18f72927ed1e8ccff9787