Group nodes by availability zone
An AZ is chosen at random from the nodepool config provider entry
when selecting a node set for a node request. If the AZs are not
listed within the config, then we cannot guarantee AZ grouping.
The one exception is if we have pre-existing READY nodes with
their AZ set. In that case, we will use the AZ from the first
READY node selected.

Change-Id: I97bdda2c4b6952f1cea18f72927ed1e8ccff9787