Refactor nodepool apps into base app
There is a lot of commonality between the nodepool apps that is
replicated between each of the files. This will mean adding more again
when new apps are added.

Seperate a base NodepoolApp and NodepoolDaemonApp that handle the common
parts of setting up a nodepool app and use them in the existing apps.

Change-Id: I098b0a1d749e9dc400d45355f9295ca07058768b