Fix min-ready/max-servers in test configs
We should never have providers in our test configs that have a
max-servers value that is LESS than the min-ready. This could
potentially cause our tests to hang (and thus fail) if such a
provider were to try and satisfy all of the min-ready requests
because that provider will pause waiting for nodes to free up.

Change-Id: I6a30cacbe97130958c2d22ff56e760549fcc6bc1