Update patch set 6
Patch Set 6:

If you end up NOT abandoning this, we need to fix the v3 config to use a different provider name, or merge the nodepool_id change, before this merges. Otherwise v3 will begin to see production nodes as leaked nodes since it would now be able to read nodepool_provider_name from the metadata and not see any ZK nodes for it.

Patch-set: 6