Update patch set 1
Patch Set 1:

recheck

FWIW, the delete failure you see in PS1 has been seen before. It would seem that kazoo's delete() code is not as thread-safe as you'd like it to be. Another lock on the node (while it's trying to delete it) will cause new child znodes to be added which are unaccounted for in the delete() call.

Patch-set: 1